Refactor AudienceSegmentationSetupCTAWidget to be a JS module.
